live-telegram/requirements.txt
pandas
numpy
matplotlib
joblib
catboost
python-telegram-bot==21.4
python-dotenv
requests